% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_param.R
\name{ess_STOCfree_param}
\alias{ess_STOCfree_param}
\title{Effective sample sizes of STOC free model parameters}
\usage{
ess_STOCfree_param(x)
}
\arguments{
\item{x}{estimated model parameters}
}
\value{

}
\description{
Effective sample sizes of STOC free model parameters
}
