% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_month.R
\name{expand_month}
\alias{expand_month}
\title{Expansion of longitudinal datasets to obtain 1 row per month}
\usage{
expand_month(data, herd_colname, date_colname, test_res_colname)
}
\arguments{
\item{data}{a dataset in the data.frame or tibble formats with 1 row per test date}

\item{herd_colname}{name of the column containing the herd identifier}

\item{date_colname}{name of the column containing the test date}

\item{test_res_colname}{name of the column containing test results}
}
\description{
Expansion of longitudinal datasets to obtain 1 row per month
}
\examples{
expand_month(data = herdBTM, herd_colname = Farm, date_colname = DateOfTest, test_res_colname = TestResult)


}
