% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{extract_herd_from_status_id}
\alias{extract_herd_from_status_id}
\title{Extract data from a STOCfree_data object based on status ids}
\usage{
extract_herd_from_status_id(sfd, status_id = integer())
}
\arguments{
\item{sfd}{a STOCfree_data object}

\item{status_id}{status id usually returned by error message}
}
\value{

}
\description{
Extract data from a STOCfree_data object based on status ids
}
\details{
This function is designed to help debugging error messages when compiling the JAGS model. Error messages can refer to Status whih is the internal variable refering to a month in a herd. The function will extract herds associated with specific statuses.
}
