% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_tidy_output.R
\name{extract_STOCfree_prevalences}
\alias{extract_STOCfree_prevalences}
\title{Extracting draws from the posterior distributions for monthly prevalences}
\usage{
extract_STOCfree_prevalences(STOCfree_model_output, STOCfree_data = NULL)
}
\arguments{
\item{STOCfree_model_output}{output from a call to STOCfree_model()}
}
\value{

}
\description{
Extracting draws from the posterior distributions for monthly prevalences
}
