% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_nwinf_lagged.R
\name{glm_nwinf}
\alias{glm_nwinf}
\title{Modelling of lagged relationships between risk factor occurrence and new infection}
\usage{
glm_nwinf(
  sf_data,
  rf_data,
  rf_herd_col = character(),
  rf_date_col = character(),
  rf_col = character(),
  lag1 = 0,
  lag2 = 36
)
}
\arguments{
\item{lag2}{}
}
\value{

}
\description{
Modelling of lagged relationships between risk factor occurrence and new infection
}
