% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_herd.R
\name{group_by_herd}
\alias{group_by_herd}
\title{Generates a dataset with herd level indices to be used by JAGS}
\usage{
group_by_herd(
  data,
  herd_colname,
  month_colname,
  row_id_colname,
  test_res_colname
)
}
\arguments{
\item{data}{a dataset with each row representing a test result in a herd on a given month}

\item{herd_colname}{name of the column with herd identifier}

\item{month_colname}{name of the column with month of test}

\item{row_id_colname}{name of the column with row identifier}

\item{test_res_colname}{name of the column with test results}
}
\value{

}
\description{
Generates a dataset with herd level indices to be used by JAGS
}
