% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_renumber.R
\name{herd_renumber}
\alias{herd_renumber}
\title{Numbering of herds from 1 to number of herds}
\usage{
herd_renumber(data, herd_colname)
}
\arguments{
\item{data}{a dataset in the data.frame or tibble formats}

\item{herd_colname}{name of the column containing the herd identifier}
}
\value{

}
\description{
Numbering of herds from 1 to number of herds
}
\examples{
herd_renumber(data = herdBTM, herd_colname = Farm)

}
