% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_mcmc_summaries.R
\name{herd_summary}
\alias{herd_summary}
\title{Summary of herd level probabilities of infection generated by the STOC free model}
\usage{
herd_summary(x = data.frame(), quantile = 0.5, cut_off = 0.5)
}
\arguments{
\item{x}{a data frame with MCMC samples for the herd level probabilities of being latent status positive}

\item{quantile}{use to summarise each herd's MCMC samples}

\item{cut_off}{cut-off value used to categorise herds as negative or positive}
}
\value{
for each herd, the quantile supplied is computed from the MCMC samples for the predicted probabilities of infection. So, if a value of 0.5 is supplied for quantile, the median probability of infection is computed. The cut_off argument is then used to categorise each herd as status positive or status negative.
}
\description{
Summary of herd level probabilities of infection generated by the STOC free model
}
