% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro.R
\docType{data}
\name{intro}
\alias{intro}
\title{Date and number of animals introduced in the tested farms}
\format{A data frame with 1100 rows and 6 variables:
\describe{
  \item{Farm}{farm ID}
  \item{dateIntr}{date of introduction}
  \item{nAnim}{number of animals}
}}
\usage{
data(herdBTM)
}
\description{
A dataset containing results of ELISA test results performed approximately
every 6 months between 2010 and 2017 in 100 herds
}
\keyword{datasets}
