% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logit function}
\usage{
invlogit(x)
}
\arguments{
\item{x}{a number}
}
\value{
a proportion
}
\description{
Inverse logit function
}
\examples{
invlogit(0)
invlogit(logit(.5)) == .5
}
