% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_nwinf_lagged.R
\name{logit_nwinf}
\alias{logit_nwinf}
\title{Logistic regression for the probability of new infection}
\usage{
logit_nwinf(nwinf = nwinf_data(), risk_factors = character)
}
\arguments{
\item{nwinf}{a dataset created with the make_nwinf_data() and add_risk_factor() functions}

\item{risk_factors}{a character string with rhe risk factors to include in the regression}
}
\value{
results of a logistic regression model performed with the glm function
}
\description{
Logistic regression for the probability of new infection
}
