% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOCfree_data.R
\name{make_animal_test}
\alias{make_animal_test}
\title{Aggregate animal level data at the herd-month level}
\usage{
make_animal_test(test_data, test_res_col, test_N_anim)
}
\arguments{
\item{test_data}{}
}
\value{

}
\description{
Aggregate animal level data at the herd-month level
}
