% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_id.R
\name{make_month_id}
\alias{make_month_id}
\title{Creates a data.frame with first date of month and corresponding month id}
\usage{
make_month_id(start = character(), end = character(), orig = start)
}
\arguments{
\item{start}{first day of the date sequence as a character string}

\item{end}{last day of the date sequence as a character string}

\item{orig}{origin of the sequence. The corresponding month is labelled 1 in the month sequence}
}
\value{
a data.frame
}
\description{
Creates a data.frame with first date of month and corresponding month id
}
\examples{
make_month_id(start = "2019-12-21", end = "2020-03-30", orig = "2020-02-20")

}
