% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_nwinf_lagged.R
\name{make_nwinf_data}
\alias{make_nwinf_data}
\title{Creates a dataset with new infection events from herd level test results}
\usage{
make_nwinf_data(sfd, time_of_inf = c("mid", "first", "last"))
}
\arguments{
\item{sfd}{a STOCfree_data object}

\item{time_of_inf}{a character string. When 'mid', the time of new infection is the median of the months between which the consecutive tests occurred.}
}
\value{

}
\description{
Creates a dataset with new infection events from herd level test results
}
