% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_output_mcmc_summaries.R
\name{param_summary}
\alias{param_summary}
\title{Summary quantiles for the STOC free model parameters}
\usage{
param_summary(
  x = data.frame(),
  parameters = c("Se", "Sp", "tau", "pi_within", "theta"),
  probs = c(0.5, 0.025, 0.975),
  digits = 4
)
}
\arguments{
\item{x}{a data frame with MCMC samples for the different model parameters generated by the STOC free model}

\item{parameters}{model parameters to summarise}

\item{probs}{quantiles to compute for each parameter. The default values are the median as well as the 2.5 and 97.5 percentiles.}

\item{digits}{number of digits used to round the results}
}
\value{
this function returns the quantiles supplied with the probs argument for the different parameters of the STOC free model.
}
\description{
Summary quantiles for the STOC free model parameters
}
