% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_month_prev.R
\name{plot.STOCfree_month_prev}
\alias{plot.STOCfree_month_prev}
\title{Boxplot of monthly prevalences predicted by the STOC free model}
\usage{
\method{plot}{STOCfree_month_prev}(month_prev)
}
\arguments{
\item{month_prev}{an object of class STOCfree_month_prev}
}
\value{

}
\description{
Boxplot of monthly prevalences predicted by the STOC free model
}
