% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_pred.R
\name{plot.STOCfree_pred}
\alias{plot.STOCfree_pred}
\title{Title}
\usage{
\method{plot}{STOCfree_pred}(x, herd = "all", type = "aggregated", legend = FALSE)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Title
}
