% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priors.R
\name{plot_priors_rf}
\alias{plot_priors_rf}
\title{Plot prior densities for the probability of new infection in the herds unexposed and exposed to the risk factors included in the STOCfree model}
\usage{
plot_priors_rf(x, n = 1e+05)
}
\arguments{
\item{x}{a list with the first element containing the mean(s) and the second element containing standard deviation(s) of normal distributions for the logistic regression}

\item{n}{number of draws from the parameter distributions used for plotting}
}
\value{
a plot of the prior distributions for the probabilities of new infection in the unexposed (intercept) and exposed to the risk factors evaluated
}
\description{
Plot prior densities for the probability of new infection in the herds unexposed and exposed to the risk factors included in the STOCfree model
}
