% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data_priors_status_dyn.R
\name{plot_priors_status_dyn}
\alias{plot_priors_status_dyn}
\title{Plots the prior distributions for the parameters related to infection dynamics}
\usage{
plot_priors_status_dyn(data)
}
\arguments{
\item{data}{a STOCfree_data object}
}
\value{

}
\description{
Plots the prior distributions for the parameters related to infection dynamics
}
