% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_output_param.R
\name{print.STOCfree_param}
\alias{print.STOCfree_param}
\title{print method for STOCfree parameters}
\usage{
\method{print}{STOCfree_param}(x)
}
\arguments{
\item{x}{an object of class STOCfree_param created with extract_STOCfree_param() or read_STOCfree_param()}
}
\value{

}
\description{
print method for STOCfree parameters
}
