% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_pred.R
\name{print.STOCfree_pred}
\alias{print.STOCfree_pred}
\title{print method for predicted probabilities of status positive}
\usage{
\method{print}{STOCfree_pred}(x)
}
\arguments{
\item{x}{an object of class STOCfree_pred}
}
\value{

}
\description{
print method for predicted probabilities of status positive
}
