% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_month_prev.R
\name{read_STOCfree_month_prev}
\alias{read_STOCfree_month_prev}
\title{Importing output files for predicted monthly prevalences generated by the STOC free model}
\usage{
read_STOCfree_month_prev(out_path = "STOCfree_files")
}
\arguments{
\item{out_path}{}
}
\value{

}
\description{
Importing output files for predicted monthly prevalences generated by the STOC free model
}
