% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_pred.R
\name{read_STOCfree_pred}
\alias{read_STOCfree_pred}
\title{Importing MCMC samples for the predicted probabilities of being latent status positive}
\usage{
read_STOCfree_pred(out_path = "STOCfree_files")
}
\arguments{
\item{out_path}{path to the folder where the results are stored}
}
\value{

}
\description{
Importing MCMC samples for the predicted probabilities of being latent status positive
}
