% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_model.R
\name{sample_model}
\alias{sample_model}
\title{Draws samples from the posterior distributions of model parameters}
\usage{
sample_model(compiled_model, n_burnin, n_iter, n_thin)
}
\arguments{
\item{compiled_model}{a model compiled by compile_JAGS()}

\item{n_burnin}{number of burnin iterations}

\item{n_iter}{number of iterations to monitor}

\item{n_thin}{thinning interval for monitors}
}
\value{
a list with 2 variables
}
\description{
Draws samples from the posterior distributions of model parameters
}
