% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_inf_dyn.R
\name{set_priors_inf_dyn}
\alias{set_priors_inf_dyn}
\title{Title}
\usage{
set_priors_inf_dyn(
  data = STOCfree_data(),
  pi1_a = 1,
  pi1_b = 1,
  tau2_a = 1,
  tau2_b = 1
)
}
\arguments{
\item{tau2_b}{}
}
\value{

}
\description{
Title
}
