% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data_priors_risk_factors.R
\name{set_priors_rf}
\alias{set_priors_rf}
\title{Set priors associated with risk factors of new infection}
\usage{
set_priors_rf(
  x = STOCfree_data(),
  risk_factor = character(),
  modality = NULL,
  mean = NULL,
  sd = NULL
)
}
\arguments{
\item{x}{an object of class STOCfree_data}

\item{risk_factor}{name of the risk factor for which priors are defined}

\item{modality}{name of the modality for which priors are defined. Only applies to categorical risk factors.}

\item{mean}{mean of the normal distribution used as a prior}

\item{sd}{standard deviation of the normal distribution used as a prior}
}
\value{

}
\description{
Set priors associated with risk factors of new infection
}
