% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_priors_tests.R
\name{set_priors_tests}
\alias{set_priors_tests}
\title{Set parameters for the Beta distributions used to represent priors on test charactersitics}
\usage{
set_priors_tests(
  x = STOCfree_data(),
  test = character(),
  Se_a = 1,
  Se_b = 1,
  Sp_a = 1,
  Sp_b = 1
)
}
\arguments{
\item{x}{a STOCfree_data object}

\item{Se_a}{alpha parameter for the prior on sensitivity}

\item{Se_b}{beta parameter for the prior on sensitivity}

\item{Sp_a}{alpha parameter for the prior on specificity}

\item{Sp_b}{beta parameter for the prior on specificity}
}
\value{

}
\description{
Set parameters for the Beta distributions used to represent priors on test charactersitics
}
