% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data.R
\name{sf_remove_risk_factor}
\alias{sf_remove_risk_factor}
\title{Remove risk factor from STOCfree_data}
\usage{
sf_remove_risk_factor(sfd, risk_factor = character())
}
\arguments{
\item{sfd}{a STOCfree_data object}

\item{risk_factor}{risk factor name}
}
\value{

}
\description{
Remove risk factor from STOCfree_data
}
