% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data_priors_status_dyn.R
\name{show_priors_status_dyn}
\alias{show_priors_status_dyn}
\title{Shows the parameters associated with priors related to infection dynamics}
\usage{
show_priors_status_dyn(data)
}
\arguments{
\item{x}{a STOC_free data object}
}
\value{
a table with the alpha and beta parameters of the Beta distributions associated with the prior probabilities of infection on the first test and the probability of remaining infected between consecutive tests
}
\description{
Shows the parameters associated with priors related to infection dynamics
}
