% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_priors_risk_factors.R
\name{show_rf}
\alias{show_rf}
\title{List of risk factors of new infection in a STOCfree_data}
\usage{
show_rf(x = STOCfree_data())
}
\arguments{
\item{x}{a STOC_free data object}
}
\value{
Displays a table with the list of risk factors considered as well as the associated priors
}
\description{
List of risk factors of new infection in a STOCfree_data
}
