% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_priors_tests.R
\name{show_tests}
\alias{show_tests}
\title{Shows parameter of the Beta distributions used to represent priors on test charactersitics}
\usage{
show_tests(x = STOCfree_data())
}
\arguments{
\item{x}{a STOCfree_data object}
}
\value{

}
\description{
Shows parameter of the Beta distributions used to represent priors on test charactersitics
}
