% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOCfree_data.R
\name{STOCfree_data}
\alias{STOCfree_data}
\title{STOCfree_data class}
\usage{
STOCfree_data(
  test_data = data.frame(),
  test_herd_col,
  test_date_col,
  test_res_col,
  test_level = c("herd", "animal"),
  risk_factor_data = NULL,
  risk_herd_col = NULL,
  risk_date_col = NULL,
  risk_factor_col = NULL,
  risk_factor_type = c("continuous", "categorical")
)
}
\arguments{
\item{test_data}{a data.frame containing test results}

\item{test_herd_col}{name of the column with herd / farm identifiers}

\item{test_date_col}{name of the column with date of test}

\item{test_res_col}{name of the column with test results}

\item{test_level}{level at which the tests are performed. Must be either herd of animal}

\item{risk_factor_data}{a data.frame containing the risk factors}

\item{risk_herd_col}{name of the column with herd / farm identifiers}

\item{risk_date_col}{name of the column with date when the risk factors apply}

\item{risk_factor_col}{name of the column(s) with risk factor values}

\item{risk_factor_type}{risk factor type. Must be either continuous or categorical}
}
\value{
A list of class STOCfree_data.
}
\description{
STOCfree_data class
}
