% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_model_JAGS_data.R
\name{STOCfree_JAGS_data}
\alias{STOCfree_JAGS_data}
\title{Creates the JAGS model input data}
\usage{
STOCfree_JAGS_data(STOCfree_data)
}
\arguments{
\item{STOCfree_data}{a STOCfree data object}
}
\description{
Creates the JAGS model input data
}
\details{
this function should only be used for checking the data
}
