% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_model.R
\name{STOCfree_model}
\alias{STOCfree_model}
\title{JAGS implementation of the STOC free model - deprecated}
\usage{
STOCfree_model(
  STOCfree_data,
  n_chains = 4,
  n_burnin = 5000,
  n_iter = 5000,
  n_thin = 20,
  method = "parallel",
  save_model = TRUE,
  save_data = FALSE,
  save_output = TRUE,
  out_path = "STOCfree_files",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
JAGS implementation of the STOC free model - deprecated
}
