% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_model.R
\name{STOCfree_model_Stan}
\alias{STOCfree_model_Stan}
\title{Stan implementation of the STOC free model - deprecated}
\usage{
STOCfree_model_Stan(
  STOCfree_data,
  n_chains = 4,
  n_iter = 1000,
  save_model = TRUE,
  save_data = FALSE,
  save_output = TRUE,
  out_path = "STOCfree_files"
)
}
\arguments{
\item{out_path}{}
}
\value{

}
\description{
Stan implementation of the STOC free model - deprecated
}
