% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_model.R
\name{STOCfree_Stan}
\alias{STOCfree_Stan}
\title{Stan implementation of the STOC free model}
\usage{
STOCfree_Stan(
  STOCfree_data,
  n_chains = 4,
  n_iter = 1000,
  n_thin = 1,
  n_warmup = NULL,
  save_model = TRUE,
  save_data = FALSE,
  save_output = TRUE,
  save_inits = TRUE,
  out_path = "STOCfree_files"
)
}
\arguments{
\item{STOCfree_data}{a STOC free data object}

\item{n_chains}{number of MCMC chains}

\item{n_iter}{number of iterations to monitor}

\item{n_thin}{thinning interval for monitors}

\item{n_warmup}{number of warmup iterations}

\item{save_model}{if TRUE, the Stan model code is saved to 'out_path'}

\item{save_data}{if TRUE, the data are saved to 'out_path' as a .RData file}

\item{save_output}{if TRUE, the JAGS model output is saved to out_path in a tidy format using the tidybayes package}

\item{save_inits}{if TRUE initial values are written to 'out_path' as a text file}

\item{out_path}{folder where model code and output are saved. By default, a STOCfree_files folder is created in the working directory}
}
\value{

}
\description{
Stan implementation of the STOC free model
}
\details{
The code used is an adaptation of Damiano et al. (2017): https://github.com/luisdamiano/stancon18
}
