% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_param.R
\name{summary.STOCfree_param}
\alias{summary.STOCfree_param}
\title{summary method for the STOCfree_param class}
\usage{
\method{summary}{STOCfree_param}(x, quantiles = c(0.025, 0.975))
}
\arguments{
\item{quantiles}{}
}
\value{

}
\description{
summary method for the STOCfree_param class
}
