% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_output_pred.R
\name{summary.STOCfree_pred}
\alias{summary.STOCfree_pred}
\title{summary method for herd-level predicted probabilities of being status positive}
\usage{
\method{summary}{STOCfree_pred}(x, herd = "all", quantiles = c(0.025, 0.975), digits = 3)
}
\arguments{
\item{x}{a STOCfree_pred object}

\item{herd}{a vector containing the herds for which a summary statistics a required}

\item{quantiles}{quantiles to be passed to the quantile() function}

\item{digits}{number of digits used in the object returned}
}
\value{

}
\description{
summary method for herd-level predicted probabilities of being status positive
}
