% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_write_Stan_model.R
\name{write_Stan_model}
\alias{write_Stan_model}
\title{Writes the code for the Stan model. For internal use.}
\usage{
write_Stan_model(data)
}
\value{
creates the Stan model
}
\description{
This function is for internal use.
}
