% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_nwinf_lagged.R
\name{add_risk_factor}
\alias{add_risk_factor}
\title{Adds a risk factor to a new infection dataset}
\usage{
add_risk_factor(
  nwinf = nwinf_data(),
  rf_data,
  rf_col = character(),
  rf_date_col = character(),
  lag1 = 6,
  lag2 = 12,
  FUN = sum
)
}
\arguments{
\item{nwinf}{a dataset created with the make_nwinf_data() function}

\item{rf_data}{risk factor data. Ids for farms must be the same as in the original test results data.}

\item{rf_col}{name of the column containing the risk factor data}

\item{rf_date_col}{name of the column with date of risk factor occurrence}

\item{lag1}{start of interval between risk factor occurrence and new infection}

\item{lag2}{end of interval between risk factor occurrence and new infection}
}
\value{

}
\description{
Adds a risk factor to a new infection dataset
}
