% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_JAGS.R
\name{compile_JAGS}
\alias{compile_JAGS}
\title{Compilation of the JAGS model}
\usage{
compile_JAGS(test_data, herd_id, row_id, month, test_result)
}
\arguments{
\item{test_data}{a tibble or a data.frame with one row per month}

\item{herd_id}{name of the column with herd identifier}

\item{row_id}{name of the column with the row IDs}

\item{month}{name of the colummn with month number}

\item{test_result}{name of the colummn with test results}
}
\value{
a compiled JAGS model
}
\description{
Compilation of the JAGS model
}
