% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_JAGS.R
\name{compile_JAGS}
\alias{compile_JAGS}
\title{Compilation of the JAGS model}
\usage{
compile_JAGS(data, n_chains = 4, keep_model_file = FALSE, ...)
}
\arguments{
\item{data}{an object of class STOCfree_data}

\item{n_chains}{number of MCMC chains}

\item{...}{}

\item{keep_model}{should the JAGS model file be saved on the disk}
}
\value{
a compiled JAGS model
}
\description{
Compilation of the JAGS model
}
