% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_JAGS.R
\name{compile_JAGS}
\alias{compile_JAGS}
\title{Compilation of the JAGS model}
\usage{
compile_JAGS(
  data,
  n_chains = 4,
  keep_model_file = FALSE,
  status = c("discrete", "proba", "predict"),
  ...
)
}
\arguments{
\item{data}{an object of class STOCfree_data}

\item{n_chains}{number of MCMC chains}

\item{keep_model_file}{if TRUE, the JAGS model is written as a text file in the working folder}

\item{status}{'discrete' latent status (0/1) is modelled, 'proba' probability of latent status is modelled, 'predict' predictive simulation from the model prior and data}
}
\value{
a compiled JAGS model
}
\description{
Compilation of the JAGS model
}
