% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_JAGS.R
\name{compile_JAGS}
\alias{compile_JAGS}
\title{Compilation of the JAGS model}
\usage{
compile_JAGS(data, n_chains = 4, ...)
}
\arguments{
\item{data}{an object of class STOCfree_data}

\item{n_chains}{number of MCMC chains}

\item{...}{}
}
\value{
a compiled JAGS model
}
\description{
Compilation of the JAGS model
}
