% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_nwinf_lagged.R
\name{date_from_lag}
\alias{date_from_lag}
\title{Generates an initial date from a reference date and a time lag in months}
\usage{
date_from_lag(date = character(), time_lag = integer())
}
\arguments{
\item{date}{a date with format yyyy-mm-dd}

\item{time_lag}{an integer representing a number of months}
}
\value{
a date which is time_lag months before or after date depending on the sign of time_lag
}
\description{
Generates an initial date from a reference date and a time lag in months
}
\examples{
seq_dates <- seq(as.Date("2020-01-01"),
 as.Date("2020-12-01"),
 by = "1 month")

data.frame(date = seq_dates,
 lagged_date = date_from_lag(
  date = sqt,
  time_lag = -2))
}
