% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_output_param.R
\name{extract_STOCfree_param}
\alias{extract_STOCfree_param}
\title{Extracting parameter MCMC samples from the results of a STOC free model}
\usage{
extract_STOCfree_param(x)
}
\arguments{
\item{x}{results returned by the STOC free model}
}
\value{

}
\description{
Extracting parameter MCMC samples from the results of a STOC free model
}
