% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_tidy_output.R
\name{extract_STOCfree_pred}
\alias{extract_STOCfree_pred}
\title{Extracting draws from the posterior distributions of predicted probabilities of infection}
\usage{
extract_STOCfree_pred(STOCfree_model_output)
}
\arguments{
\item{STOCfree_model_output}{output from a call to STOCfree_model()}
}
\value{

}
\description{
Extracting draws from the posterior distributions of predicted probabilities of infection
}
