% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_output_pred.R
\name{extract_STOCfree_pred}
\alias{extract_STOCfree_pred}
\title{Extracting MCMC samples for the predicted probabilities of being latent status positive from the results of a STOC free model}
\usage{
extract_STOCfree_pred(x, STOCfree_data = NULL)
}
\arguments{
\item{x}{results returned by the STOC free model}
}
\value{

}
\description{
Extracting MCMC samples for the predicted probabilities of being latent status positive from the results of a STOC free model
}
