% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_data_priors_risk_factors.R
\name{plot_priors_rf}
\alias{plot_priors_rf}
\title{Plotting prior distributions for the risk factors of new infection}
\usage{
plot_priors_rf(x = STOCfree_data(), n = 1e+05)
}
\arguments{
\item{x}{a dataset in the STOCfree_data format}

\item{n}{number of samples to draw from the prior distribution}
}
\value{

}
\description{
Plotting prior distributions for the risk factors of new infection
}
