% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_tests.R
\name{plot_priors_tests}
\alias{plot_priors_tests}
\title{Plot the distribution for the prior distributions on test characteristics}
\usage{
plot_priors_tests(data = STOCfree_data())
}
\arguments{
\item{x}{a STOCfree_data object}
}
\value{

}
\description{
Plot the distribution for the prior distributions on test characteristics
}
