% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_param.R
\name{read_STOCfree_param}
\alias{read_STOCfree_param}
\title{Importing MCMC samples from the results of a STOC free model}
\usage{
read_STOCfree_param(out_path = "STOCfree_files")
}
\arguments{
\item{out_path}{}
}
\value{

}
\description{
Importing MCMC samples from the results of a STOC free model
}
