% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_output_param.R
\name{read_STOCfree_param}
\alias{read_STOCfree_param}
\title{Importing MCMC samples for model parameters from the results of a STOC free model}
\usage{
read_STOCfree_param(out_path = "STOCfree_files")
}
\arguments{
\item{out_path}{folder in which the model, data and result files will be stored}
}
\value{

}
\description{
Importing MCMC samples for model parameters from the results of a STOC free model
}
