% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_tidy_output.R
\name{read_STOCfree_pred}
\alias{read_STOCfree_pred}
\title{Read output files for predicted probabilities of infection generated by the STOC free model}
\usage{
read_STOCfree_pred(out_path = "STOCfree_files")
}
\arguments{
\item{out_path}{folder where model outputs are saved. By default, a STOCfree_files folder is created in the working directory}
}
\value{

}
\description{
Read output files for predicted probabilities of infection generated by the STOC free model
}
