% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_priors_inf_dyn.R
\name{set_priors_inf_dyn}
\alias{set_priors_inf_dyn}
\title{Sets the parameters associated with priors related to infection dynamics}
\usage{
set_priors_inf_dyn(
  data = NULL,
  pi1_a = NULL,
  pi1_b = NULL,
  logit_pi1_mean = NULL,
  logit_pi1_sd = NULL,
  tau1_a = NULL,
  tau1_b = NULL,
  logit_tau1_mean = NULL,
  logit_tau1_sd = NULL,
  tau2_a = NULL,
  tau2_b = NULL,
  logit_tau2_mean = NULL,
  logit_tau2_sd = NULL,
  pi_within_a = NULL,
  pi_within_b = NULL,
  logit_pi_within_mean = NULL,
  logit_pi_within_sd = NULL
)
}
\arguments{
\item{data}{a STOCfree_data object}

\item{pi1_a}{alpha parameter for the prior distribution of the probability of being infected on the first test}

\item{pi1_b}{beta parameter for the prior distribution of the probability of being infected on the first test}

\item{tau2_a}{alpha parameter for the prior distribution of the probability of remaining infected between consecutive months}

\item{tau2_b}{beta parameter for the prior distribution of the probability of remaining infected between consecutive months}
}
\value{

}
\description{
Sets the parameters associated with priors related to infection dynamics
}
