% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_inf_dyn.R
\name{set_priors_inf_dyn}
\alias{set_priors_inf_dyn}
\title{Sets the parameters asssociated with priors related to infection dynamics}
\usage{
set_priors_inf_dyn(
  data = STOCfree_data(),
  pi1_a = 1,
  pi1_b = 1,
  pi_within_a = 1,
  pi_within_b = 1,
  tau1_a = 1,
  tau1_b = 1,
  tau2_a = 1,
  tau2_b = 1
)
}
\arguments{
\item{data}{a STOCfree_data object}

\item{pi1_a}{alpha parameter for the prior distribution of the probability of being infected on the first test}

\item{pi1_b}{beta parameter for the prior distribution of the probability of being infected on the first test}

\item{tau2_a}{alpha parameter for the prior distribution of the probability of remaining infected between consecutive months}

\item{tau2_b}{beta parameter for the prior distribution of the probability of remaining infected between consecutive months}
}
\value{

}
\description{
Sets the parameters asssociated with priors related to infection dynamics
}
