% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOCfree_data.R
\name{sf_add_risk_factor}
\alias{sf_add_risk_factor}
\title{Add a risk factor to STOCfree_data}
\usage{
sf_add_risk_factor(
  sfd,
  risk_factor_data,
  risk_herd_col = NULL,
  risk_date_col = NULL,
  risk_factor_col = NULL,
  risk_factor_type = c("continuous", "categorical"),
  lag1 = 0,
  lag2 = 0,
  FUN = sum
)
}
\arguments{
\item{sfd}{STOC free data}

\item{risk_factor_data}{a data.frame with a risk factor}

\item{FUN}{function used between lag1 and lag2. The default is a sum.}
}
\value{

}
\description{
Add a risk factor to STOCfree_data
}
