% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOCfree_data.R
\name{STOCfree_data}
\alias{STOCfree_data}
\title{STOCfree_data class}
\usage{
STOCfree_data(
  test_data = data.frame(),
  test_herd_col = NULL,
  test_date_col = NULL,
  test_res_col = NULL,
  test_name_col = NULL,
  test_level = c("herd", "animal"),
  test_N_anim = NULL,
  risk_factor_data = NULL,
  risk_herd_col = NULL,
  risk_date_col = NULL,
  risk_factor_col = NULL,
  risk_factor_type = c("continuous", "categorical"),
  lag1 = 0,
  lag2 = 0,
  FUN = sum
)
}
\arguments{
\item{test_data}{a data.frame containing test results}

\item{test_herd_col}{name of the column with herd / farm identifiers}

\item{test_date_col}{name of the column with date of test}

\item{test_res_col}{name of the column with test results}

\item{test_name_col}{when several tests are used, name of the column containing the test names}

\item{test_level}{level at which the tests are performed. Must be either herd of animal}

\item{test_N_anim}{name of the column with the number of animals to use as the denominator in the animal level model.}

\item{risk_factor_data}{a data.frame containing the risk factors}

\item{risk_herd_col}{name of the column with herd / farm identifiers}

\item{risk_date_col}{name of the column with date when the risk factors apply}

\item{risk_factor_col}{name of the column(s) with risk factor values}

\item{risk_factor_type}{risk factor type. Must be either continuous or categorical}

\item{FUN}{}
}
\value{
A list of class STOCfree_data. A specific sub-class is defined based on the input.
}
\description{
STOCfree_data class
}
