% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_model.R
\name{STOCfree_model}
\alias{STOCfree_model}
\title{Draws samples from the posterior distributions of model parameters}
\usage{
STOCfree_model(
  STOCfree_data,
  n_chains = 4,
  n_burnin = 5000,
  n_iter = 5000,
  n_thin = 20,
  method = "parallel",
  save_model = TRUE,
  save_output = TRUE,
  out_path = "STOCfree_files",
  ...
)
}
\arguments{
\item{STOCfree_data}{a STOC free data object}

\item{n_chains}{number of MCMC chains}

\item{n_burnin}{number of burnin iterations}

\item{n_iter}{number of iterations to monitor}

\item{n_thin}{thinning interval for monitors}

\item{method}{method to be passed to the runjags::run.jags() function. Default is parallel.}

\item{save_model}{if TRUE, the JAGS model code is saved to out_path}

\item{save_output}{if TRUE, the JAGS model output is saved to out_path in a tidy format using the tidybayes package}

\item{out_path}{folder where model code and output are saved. By default, a STOCfree_files folder is created in the working directory}

\item{...}{}
}
\description{
Draws samples from the posterior distributions of model parameters
}
