% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_model.R
\name{STOCfree_model_Stan}
\alias{STOCfree_model_Stan}
\title{Draws samples from the posterior distributions of model parameters using Stan}
\usage{
STOCfree_model_Stan(
  STOCfree_data,
  n_chains = 4,
  n_iter = 1000,
  save_output = TRUE,
  out_path = "STOCfree_Stan_files"
)
}
\arguments{
\item{STOCfree_data}{a STOC free data object}

\item{save_output}{if TRUE, the JAGS model output is saved to out_path in a tidy format using the tidybayes package}

\item{out_path}{folder where model code and output are saved. By default, a STOCfree_files folder is created in the working directory}
}
\value{

}
\description{
Draws samples from the posterior distributions of model parameters using Stan
}
\details{
The code used is an adaptation of Damiano et al. (2017): https://github.com/luisdamiano/stancon18
}
