% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_output_param.R
\name{summary.STOCfree_param}
\alias{summary.STOCfree_param}
\title{summary method for the STOCfree_param class}
\usage{
\method{summary}{STOCfree_param}(x, quantiles = c(0.025, 0.975))
}
\arguments{
\item{x}{an object of class STOCfree_param created with extract_STOCfree_param() or read_STOCfree_param()}

\item{quantiles}{quantiles of the posterior distributions to display. Default values are 0.025 and 0.975}
}
\value{

}
\description{
summary method for the STOCfree_param class
}
