% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_model_JAGS_write_model.R
\name{write_JAGS_model}
\alias{write_JAGS_model}
\title{Writes a text file with the JAGS model. For internal use.}
\usage{
write_JAGS_model(data)
}
\value{
a text file with the JAGS model
}
\description{
This function is for internal use.
}
